CI and Extension Developer Cleanup

+++ scripts/ci_install.sh
@@ -7,14 +7,16 @@ set -o pipefail
 
 # Building should work without yarn installed globally, so uninstall the
 # global yarn installed by default.
-sudo rm -rf $(which yarn)
-! yarn
+if [ $OSTYPE == "Linux" ]; then
+    sudo rm -rf $(which yarn)
+    ! yarn
+fi
 
 # create jupyter base dir (needed for config retrieval)
 mkdir ~/.jupyter
 
 # Install and enable the server extension
-pip install -q --upgrade pip
+pip install -q --upgrade pip --user
 pip --version
 # Show a verbose install if the install fails, for debugging
 pip install -e ".[test]" || pip install -v -e ".[test]"
